[Fix] BaseFriendCell에 기존 Components 삭제 및 FavorProfilePhotoView 추가
